+++ spec/support/shared_examples.rb
@@ -1,0 +1,11 @@
+RSpec.shared_examples 'capture stdout' do
+  before do
+    require 'stringio'
+    $stdout = StringIO.new
+    @output = $stdout
+  end
+
+  after do
+    $stdout = STDOUT
+  end
+end